Double-precision camera changes
Script sync from SHA: 9708ab3152eec14cdc12b6a49fb6020369ed0d1a

+++ src/animatedclusters_ui.cpp
@@ -1,5 +1,5 @@
 /*
-* Copyright (c) 2024-2025, NVIDIA CORPORATION.  All rights reserved.
+* Copyright (c) 2024-2026, NVIDIA CORPORATION.  All rights reserved.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
-* SPDX-FileCopyrightText: Copyright (c) 2024-2025, NVIDIA CORPORATION.
+* SPDX-FileCopyrightText: Copyright (c) 2024-2026, NVIDIA CORPORATION.
 * SPDX-License-Identifier: Apache-2.0
 */
 
@@ -166,7 +166,7 @@ void AnimatedClusters::onUIRender()
       const glm::vec3 hitPos   = glm::unProjectZO({mousePos.x, mousePos.y, d}, view, proj, win_norm);
 
       // Set the interest position
-      glm::vec3 eye, center, up;
+      glm::dvec3 eye, center, up;
       m_info.cameraManipulator->getLookat(eye, center, up);
       m_info.cameraManipulator->setLookat(eye, hitPos, up, false);
     }
